+++ CHANGELOG.md
@@ -12,3 +12,6 @@
 ### 0.9.3
 ### 0.9.4
 * Minor UI changed
+### 0.9.5
+* Added output and log colorization
+* Fixed virtual terminal - related bug